SquishyNavbar: If power saver is on, don't allow feedback
When power saver is on, the squeeze vibration is turned off. This is done
by Google to save battery life across the system. Keeping the navbar animation
on doesn't make sense if we're trying to save power. This commit fixes that.

- Also, make sure we're not calling the navbar animation while the screen
is turned off. Although we don't see it, we're technically still calling on
it on.

Change-Id: I94d416b0227d68885cc07422f43a84ca70b1d26d
Signed-off-by: Josh Fox (XlxFoXxlX) <joshfox87@gmail.com>